internal/jujuapi: refactor facades
Move each facade into its own file to make the implementations easier
to find.
