+++ internal/jujuapi/admin_test.go
@@ -1,4 +1,4 @@
-// Copyright 2017 Canonical Ltd.
+// Copyright 2016 Canonical Ltd.
 
 package jujuapi_test
 
@@ -8,31 +8,101 @@ import (
 	"github.com/juju/juju/api"
 	jujuparams "github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/core/network"
+	"github.com/juju/juju/rpc"
 	"github.com/juju/names/v4"
 	jc "github.com/juju/testing/checkers"
+	"github.com/juju/utils"
 	gc "gopkg.in/check.v1"
-	"gopkg.in/errgo.v1"
+	errgo "gopkg.in/errgo.v1"
+	"gopkg.in/macaroon.v2"
 
 	"github.com/CanonicalLtd/jimm/params"
 )
 
-type modelSuite struct {
+type adminSuite struct {
 	websocketSuite
 }
 
-var _ = gc.Suite(&modelSuite{})
+var _ = gc.Suite(&adminSuite{})
 
-func (s *modelSuite) TestUnknownModel(c *gc.C) {
-	conn := s.open(c, &api.Info{
-		ModelTag:  names.NewModelTag("00000000-0000-0000-0000-000000000000"),
-		SkipLogin: true,
-	}, "bob")
-	defer conn.Close()
-	err := conn.Login(nil, "", "", nil)
-	c.Assert(err, gc.ErrorMatches, `model "00000000-0000-0000-0000-000000000000" not found \(model not found\)`)
+func (s *adminSuite) SetUpTest(c *gc.C) {
+	s.ServerParams.CharmstoreLocation = "https://api.jujucharms.com/charmstore"
+	s.ServerParams.MeteringLocation = "https://api.jujucharms.com/omnibus"
+	s.websocketSuite.SetUpTest(c)
+	s.PatchValue(&utils.OutgoingAccessAllowed, true)
 }
 
-func (s *modelSuite) TestLoginToModel(c *gc.C) {
+func (s *adminSuite) TestOldAdminVersionFails(c *gc.C) {
+	ctx := context.Background()
+
+	s.AssertAddController(ctx, c, params.EntityPath{User: "test", Name: "controller-1"}, true)
+	cred := s.AssertUpdateCredential(ctx, c, "test", "dummy", "cred1", "empty")
+	mi := s.assertCreateModel(c, createModelParams{name: "model-1", username: "test", cred: cred})
+	modelUUID := mi.UUID
+	conn := s.open(c, &api.Info{
+		ModelTag:  names.NewModelTag(modelUUID),
+		SkipLogin: true,
+	}, "test")
+	defer conn.Close()
+	var resp jujuparams.RedirectInfoResult
+	err := conn.APICall("Admin", 2, "", "Login", nil, &resp)
+	c.Assert(err, gc.ErrorMatches, `JIMM does not support login from old clients \(not supported\)`)
+	c.Assert(resp, jc.DeepEquals, jujuparams.RedirectInfoResult{})
+}
+
+func (s *adminSuite) TestAdminIDFails(c *gc.C) {
+	ctx := context.Background()
+
+	s.AssertAddController(ctx, c, params.EntityPath{User: "test", Name: "controller-1"}, true)
+	cred := s.AssertUpdateCredential(ctx, c, "test", "dummy", "cred1", "empty")
+	mi := s.assertCreateModel(c, createModelParams{name: "model-1", username: "test", cred: cred})
+	modelUUID := mi.UUID
+	conn := s.open(c, &api.Info{
+		ModelTag:  names.NewModelTag(modelUUID),
+		SkipLogin: true,
+	}, "test")
+	defer conn.Close()
+	var resp jujuparams.RedirectInfoResult
+	err := conn.APICall("Admin", 3, "Object ID", "Login", nil, &resp)
+	c.Assert(err, gc.ErrorMatches, "id not found")
+}
+
+func (s *adminSuite) TestLoginToController(c *gc.C) {
+	ctx := context.Background()
+
+	s.AssertAddController(ctx, c, params.EntityPath{User: "test", Name: "controller-1"}, true)
+	conn := s.open(c, &api.Info{
+		SkipLogin: true,
+	}, "test")
+	defer conn.Close()
+	err := conn.Login(nil, "", "", nil)
+	c.Assert(err, gc.Equals, nil)
+	var resp jujuparams.RedirectInfoResult
+	err = conn.APICall("Admin", 3, "", "RedirectInfo", nil, &resp)
+	rerr, ok := errgo.Cause(err).(*rpc.RequestError)
+	c.Assert(ok, gc.Equals, true)
+	c.Assert(rerr.Code, gc.Equals, jujuparams.CodeNotImplemented)
+}
+
+func (s *adminSuite) TestLoginToControllerWithInvalidMacaroon(c *gc.C) {
+	ctx := context.Background()
+
+	s.AssertAddController(ctx, c, params.EntityPath{User: "test", Name: "controller-1"}, true)
+	invalidMacaroon, err := macaroon.New(nil, []byte("invalid"), "", macaroon.V1)
+	c.Assert(err, gc.Equals, nil)
+	conn := s.open(c, &api.Info{
+		Macaroons: []macaroon.Slice{{invalidMacaroon}},
+	}, "test")
+	conn.Close()
+}
+
+type modelAdminSuite struct {
+	websocketSuite
+}
+
+var _ = gc.Suite(&modelAdminSuite{})
+
+func (s *modelAdminSuite) TestLoginToModel(c *gc.C) {
 	ctx := context.Background()
 
 	s.AssertAddController(ctx, c, params.EntityPath{User: "test", Name: "controller-1"}, true)
@@ -65,7 +135,7 @@ func (s *modelSuite) TestLoginToModel(c *gc.C) {
 	})
 }
 
-func (s *modelSuite) TestOldAdminVersionFails(c *gc.C) {
+func (s *modelAdminSuite) TestOldAdminVersionFails(c *gc.C) {
 	ctx := context.Background()
 
 	s.AssertAddController(ctx, c, params.EntityPath{User: "test", Name: "controller-1"}, true)
@@ -83,7 +153,7 @@ func (s *modelSuite) TestOldAdminVersionFails(c *gc.C) {
 	c.Assert(resp, jc.DeepEquals, jujuparams.RedirectInfoResult{})
 }
 
-func (s *modelSuite) TestAdminIDFails(c *gc.C) {
+func (s *modelAdminSuite) TestAdminIDFails(c *gc.C) {
 	ctx := context.Background()
 
 	s.AssertAddController(ctx, c, params.EntityPath{User: "test", Name: "controller-1"}, true)
@@ -100,7 +170,7 @@ func (s *modelSuite) TestAdminIDFails(c *gc.C) {
 	c.Assert(err, gc.ErrorMatches, "id not found")
 }
 
-func (s *modelSuite) TestUnimplementedMethodFails(c *gc.C) {
+func (s *modelAdminSuite) TestUnimplementedMethodFails(c *gc.C) {
 	ctx := context.Background()
 
 	s.AssertAddController(ctx, c, params.EntityPath{User: "test", Name: "controller-1"}, true)
@@ -116,7 +186,7 @@ func (s *modelSuite) TestUnimplementedMethodFails(c *gc.C) {
 	c.Assert(err, gc.ErrorMatches, `no such request - method Admin.Logout is not implemented \(not implemented\)`)
 }
 
-func (s *modelSuite) TestUnimplementedRootFails(c *gc.C) {
+func (s *modelAdminSuite) TestUnimplementedRootFails(c *gc.C) {
 	ctx := context.Background()
 
 	s.AssertAddController(ctx, c, params.EntityPath{User: "test", Name: "controller-1"}, true)
