+++ internal/jujuapi/usermanager_test.go
@@ -1,0 +1,139 @@
+// Copyright 2016 Canonical Ltd.
+
+package jujuapi_test
+
+import (
+	"github.com/juju/juju/api/usermanager"
+	jujuparams "github.com/juju/juju/apiserver/params"
+	jc "github.com/juju/testing/checkers"
+	"github.com/juju/utils"
+	gc "gopkg.in/check.v1"
+)
+
+type usermanagerSuite struct {
+	websocketSuite
+}
+
+var _ = gc.Suite(&usermanagerSuite{})
+
+func (s *usermanagerSuite) SetUpTest(c *gc.C) {
+	s.ServerParams.CharmstoreLocation = "https://api.jujucharms.com/charmstore"
+	s.ServerParams.MeteringLocation = "https://api.jujucharms.com/omnibus"
+	s.websocketSuite.SetUpTest(c)
+	s.PatchValue(&utils.OutgoingAccessAllowed, true)
+}
+
+func (s *usermanagerSuite) TestAddUser(c *gc.C) {
+	conn := s.open(c, nil, "alice")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	_, _, err := client.AddUser("bob", "Bob", "bob's super secret password")
+	c.Assert(err, gc.ErrorMatches, `unauthorized \(unauthorized access\)`)
+}
+
+func (s *usermanagerSuite) TestRemoveUser(c *gc.C) {
+	conn := s.open(c, nil, "alice")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	err := client.RemoveUser("bob")
+	c.Assert(err, gc.ErrorMatches, `unauthorized \(unauthorized access\)`)
+}
+
+func (s *usermanagerSuite) TestEnableUser(c *gc.C) {
+	conn := s.open(c, nil, "alice")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	err := client.EnableUser("bob")
+	c.Assert(err, gc.ErrorMatches, `unauthorized \(unauthorized access\)`)
+}
+
+func (s *usermanagerSuite) TestDisableUser(c *gc.C) {
+	conn := s.open(c, nil, "alice")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	err := client.DisableUser("bob")
+	c.Assert(err, gc.ErrorMatches, `unauthorized \(unauthorized access\)`)
+}
+
+func (s *usermanagerSuite) TestUserInfoAllUsers(c *gc.C) {
+	conn := s.open(c, nil, "alice")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	users, err := client.UserInfo(nil, usermanager.AllUsers)
+	c.Assert(err, gc.Equals, nil)
+	c.Assert(len(users), gc.Equals, 0)
+}
+
+func (s *usermanagerSuite) TestUserInfoSpecifiedUser(c *gc.C) {
+	conn := s.open(c, nil, "alice")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	users, err := client.UserInfo([]string{"alice@external"}, usermanager.AllUsers)
+	c.Assert(err, gc.Equals, nil)
+	c.Assert(len(users), gc.Equals, 1)
+	c.Assert(users[0], jc.DeepEquals, jujuparams.UserInfo{
+		Username:    "alice@external",
+		DisplayName: "alice@external",
+		Access:      "add-model",
+	})
+}
+
+func (s *usermanagerSuite) TestUserInfoSpecifiedUsers(c *gc.C) {
+	conn := s.open(c, nil, "alice")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	users, err := client.UserInfo([]string{"alice@external", "bob@external"}, usermanager.AllUsers)
+	c.Assert(err, gc.ErrorMatches, "bob@external: unauthorized")
+	c.Assert(users, gc.HasLen, 0)
+}
+
+func (s *usermanagerSuite) TestUserInfoWithDomain(c *gc.C) {
+	conn := s.open(c, nil, "alice@mydomain")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	users, err := client.UserInfo([]string{"alice@mydomain"}, usermanager.AllUsers)
+	c.Assert(err, gc.Equals, nil)
+	c.Assert(len(users), gc.Equals, 1)
+	c.Assert(users[0], jc.DeepEquals, jujuparams.UserInfo{
+		Username:    "alice@mydomain",
+		DisplayName: "alice@mydomain",
+		Access:      "add-model",
+	})
+}
+
+func (s *usermanagerSuite) TestUserInfoInvalidUsername(c *gc.C) {
+	conn := s.open(c, nil, "alice")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	users, err := client.UserInfo([]string{"alice-@external"}, usermanager.AllUsers)
+	c.Assert(err, gc.ErrorMatches, `"alice-@external" is not a valid username`)
+	c.Assert(users, gc.HasLen, 0)
+}
+
+func (s *usermanagerSuite) TestUserInfoLocalUsername(c *gc.C) {
+	conn := s.open(c, nil, "alice")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	users, err := client.UserInfo([]string{"alice"}, usermanager.AllUsers)
+	c.Assert(err, gc.ErrorMatches, `alice: unsupported local user`)
+	c.Assert(users, gc.HasLen, 0)
+}
+
+func (s *usermanagerSuite) TestSetPassword(c *gc.C) {
+	conn := s.open(c, nil, "alice")
+	defer conn.Close()
+
+	client := usermanager.NewClient(conn)
+	err := client.SetPassword("bob", "bob's new super secret password")
+	c.Assert(err, gc.ErrorMatches, `unauthorized \(unauthorized access\)`)
+}
